restructured directoy
