+++ dweet2ser/webapp/__init__.py
@@ -1,0 +1,6 @@
+import flask
+import flask_socketio
+
+webapp = flask.Flask(__name__)
+webapp.app_context().push()
+socketio = flask_socketio.SocketIO(webapp)
